third_party/WebKit/LayoutTests/bluetooth/requestDevice/acceptAllDevices/optional-services-present.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return setBluetoothFakeAdapter('TwoHeartRateServicesAdapter')
    .then(() => requestDeviceWithTrustedClick({
      acceptAllDevices: true,
      optionalServices: ['heart_rate']
    }))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryServices())
    .then(services => {
      assert_equals(services.length, 2);
      services.forEach(service => {
        assert_equals(service.uuid, BluetoothUUID.getService('heart_rate'));
      });
    })
}, 'requestDevice called with acceptAllDevices: true and with ' +
   'optionalServices. Should get access to services.');
</script>
